Exception handling

+++ src/main/java/com/spring/boot/ecommerce/services/ProductService.java
@@ -1,11 +1,17 @@
 package com.spring.boot.ecommerce.services;
 
+import com.fasterxml.jackson.databind.exc.InvalidDefinitionException;
 import com.spring.boot.ecommerce.dtos.ProductDTO;
 import com.spring.boot.ecommerce.entities.Product;
 import com.spring.boot.ecommerce.repositories.ProductRepository;
+import com.spring.boot.ecommerce.services.exceptions.DatabaseException;
+import com.spring.boot.ecommerce.services.exceptions.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
@@ -19,7 +25,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id) {
-        Product product = repository.findById(id).get();
+        Product product = repository.findById(id).
+                orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado!"));
         return new ProductDTO(product);
     }
 
@@ -39,16 +46,29 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(product, dto);
+        try{
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(product, dto);
 
-        product = repository.save(product);
-        return new ProductDTO(product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        }
+        catch(EntityNotFoundException e){
+            throw new ResourceNotFoundException("Produto não encontrado.");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void deleteById(Long id){
-        repository.deleteById(id);
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Produto não encontrado.");
+        }
+        try {
+            repository.deleteById(id);
+        }
+        catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial.");
+        }
     }
 
     private void copyDtoToEntity(Product product, ProductDTO dto) {
